Enable remote updates
Automatically checked daily; also possible to trigger from the app menu.

It turns out that Mac apps must be code-signed for Electron's auto-updater to
work, so we must disable the ability for the server to build the app (since
only Macs can code-sign and it is unlikely that the server is a Mac).